Tex-Structure Update
Tex Structure has been unified for all assignments. Each assignment is to be written into separate tex file, that have been created in the TexStructure folder.  Any assignment between 1 and 14 can be binded into master file by changing the assignment number in the \addAssignment function in Main.tex. The submission date is also updated with respect to the included assignment.